requirements.txt
shopifyapi
streamlit
python-dotenv